+++ src/lib/components/directive-collection.ts
@@ -1,7 +1,6 @@
 import { TerraDataTableContextMenuDirective } from './tables/data-table/context-menu/terra-data-table-context-menu.directive';
 import { Type } from '@angular/core';
 import { TerraTwoColumnsContainerDirective } from './layouts/column-container/two-columns/terra-two-columns-container.directive';
-import { TerraLabelTooltipDirective } from '../helpers/terra-label-tooltip.directive';
 import { FormEntryContainerDirective } from './forms/form/form-entry/form-entry-container.directive';
 import { CKEditorDirective } from './editors/ck-editor/ck-editor.directive';
 import { TooltipDirective } from './tooltip/tooltip.directive';
@@ -10,7 +9,6 @@ import { SelectSortDirective } from './forms/select-box/directive/select-sort.di
 export const directives: Array<Type<any>> = [
     TerraDataTableContextMenuDirective,
     TerraTwoColumnsContainerDirective,
-    TerraLabelTooltipDirective,
     FormEntryContainerDirective,
     CKEditorDirective,
     TooltipDirective,